+++ libclamav/others.h
@@ -44,11 +44,11 @@ extern uint8_t cli_debug_flag, cli_leavetemps_flag;
  * The macro can be used to protect against wraps.
  */
 #define CLI_ISCONTAINED(bb, bb_size, sb, sb_size)	\
-    (bb_size > 0 && sb_size > 0 && sb_size <= bb_size	\
+    (bb_size > 0 && sb_size > 0 && (size_t)sb_size <= (size_t)bb_size	\
      && sb >= bb && sb + sb_size <= bb + bb_size && sb + sb_size > bb)
 
 #define CLI_ISCONTAINED2(bb, bb_size, sb, sb_size)	\
-    (bb_size > 0 && sb_size >= 0 && sb_size <= bb_size	\
+    (bb_size > 0 && sb_size >= 0 && (size_t)sb_size <= (size_t)bb_size	\
      && sb >= bb && sb + sb_size <= bb + bb_size && sb + sb_size >= bb)
 
 #define CLI_MAX_ALLOCATION 184549376
